requirements.txt
asgiref==3.8.1
Django~=5.2
Pillow~=11.2
sqlparse==0.5.0
django-braces==1.15.0
django-embed-video==1.4.9
pymemcache==4.0.0
django-debug-toolbar==4.3.0
redis==5.0.4
django-redisboard==8.4.0
djangorestframework==3.15.1
requests==2.31.0
channels[daphne]==4.1.0
channels-redis==4.2.0
psycopg[binary]==3.1.18
uwsgi==2.0.25.1
python-decouple==3.8